requirements.txt
dbt-clickhouse>=1.8.0
pandas
apache-airflow-providers-fab